FujiDelegates Bugfix
Corrected a bug where the Fujifilm delegate would only work for JPG by forgetting to take the Exif Image Width and Image Height into account in the calculation.The Focus Pixel argument is indicated relative to the width and height. As the raf only contains a 1920w image, the Focus Pixel argument is calculated accordingly.