Draw af point with crop and rotation
There is a small amount of error still in the algorithm that gets
amplified with heavy cropping.  I feel like its in the scaling stage.
Still, this works well for most cases and can be improved upon slowly.

The focus box needs to be rotated to match the photo rotation, so for
now the upper left corners are connected.

Tested with crop, rotation, crop + rotation, offset crop, offset crop +
rotation on horizontal and vertical images.