Starting migration to center anchored focus points
Modified DefaultPointRenderer to assume that the focusPoints are now
the center of the points and not the left top corner anymore.
Temporarily patched the DefaultDelegate to artificially add back half
of the focus point dimensions until the config files have been updated.
Designed a new focus_point image to better accommodate the fact that
the picture can not be resized to the right size of the focus point.