Add rough support for Sony
This work is based on files from an a7r3, so it may not work with other
models. The values used for calculations are guesses based on marketing
material from Sony and trial and error. The intended features are the
following.

* Show the focus point that was used
* Show the outer frame of the PDAF area
* Show active PDAF locations

Most Sony cameras that have PDAF only use it down to f/8 and only over a
subset of the sensor, so the lack of related EXIF data should not be an
error. No attempt was made to support eye/face detection, nor to
identify the specific focus area used for a given image.

Related to musselwhizzle/Focus-Points#120