Olympus/OM: revert to center dot focus point
Issue #144 and related fix was nonsense.
For Olympus/OM cameras, the only useful information related to focus point is AFPointSelected. Drawing a box around this point has no added meaning in terms of focusing, but it helps to recognize / find the point more easily on the image.
Reverting implementation to POINTTYPE_AF_SELECTED_INFOCUS.